+++ extensions/wrapper/wrapper/src/main/java/de/sovity/edc/ext/wrapper/WrapperExtensionContextBuilder.java
@@ -30,6 +30,7 @@ import de.sovity.edc.ext.wrapper.api.common.mappers.utils.UiAssetMapper;
 import de.sovity.edc.ext.wrapper.api.ui.UiResourceImpl;
 import de.sovity.edc.ext.wrapper.api.ui.pages.asset.AssetApiService;
 import de.sovity.edc.ext.wrapper.api.ui.pages.asset.AssetIdValidator;
+import de.sovity.edc.ext.wrapper.api.ui.pages.asset.AssetUpdater;
 import de.sovity.edc.ext.wrapper.api.ui.pages.catalog.CatalogApiService;
 import de.sovity.edc.ext.wrapper.api.ui.pages.contract_agreements.ContractAgreementPageApiService;
 import de.sovity.edc.ext.wrapper.api.ui.pages.contract_agreements.ContractAgreementTransferApiService;
@@ -177,10 +178,11 @@ public class WrapperExtensionContextBuilder {
         );
         var contractAgreementUtils = new ContractAgreementUtils(contractAgreementService);
         var assetIdValidator = new AssetIdValidator();
+        var assetUpdater = new AssetUpdater(assetMapper, assetIdValidator, selfDescriptionService);
         var assetApiService = new AssetApiService(
                 assetService,
                 assetMapper,
-                assetIdValidator,
+                assetUpdater,
                 selfDescriptionService
         );
         var transferRequestBuilder = new TransferRequestBuilder(